src/requirements.txt
numpy
pandas
matplotlib
seaborn
statsmodels
openai
python-dotenv